requirements.txt
# langchain
# langchain_community
# PyMuPDF

# google
# langchain_google_genai==2.1.9

# pymilvus
# milvus
# fastapi

# uvicorn
# gunicorn
# google-cloud-translate
# google-cloud-storage
# langdetect
# google-cloud-discoveryengine
# google-cloud-aiplatform


# pypdf
# pillow



# LangChain Core and Community
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Google AI Integration (Use this instead of generic 'google')
langchain-google-genai==2.1.9
google-cloud-translate
google-cloud-storage
google-cloud-discoveryengine
google-cloud-aiplatform

# Database
pymilvus

# PDF and Image Processing
PyMuPDF
pypdf
pillow

# API and Server
fastapi
uvicorn
gunicorn
pydantic
python-dotenv
PyYAML

# Language Detection
langdetect